refactor: useSourceActions splitted in 3 hooks;
