+++ packages/shared/src/hooks/source/index.ts
@@ -1,2 +1,2 @@
 export * from './useSourceActions';
-export * from './useSourceSubscription';
+export * from './useSourceActionsNotify';
